Use float for calculations for compatibility with Vec2 and Vec3

+++ src/flappy/utilities/Rotation.hpp
@@ -4,31 +4,31 @@
 
 namespace flappy { namespace rotation {
 
-inline double distance(double x, double y) {
+inline float distance(float x, float y) {
     return sqrt(x * x + y * y);
 }
 
-inline double roll(cocos2d::Vec3 point) {
+inline float roll(cocos2d::Vec3 point) {
     return atan(point.x / (distance(point.y, point.z) + MATH_FLOAT_SMALL));
 }
 
-inline double pitch(cocos2d::Vec3 point) {
+inline float pitch(cocos2d::Vec3 point) {
     return atan(point.y / (distance(point.x, point.z) + MATH_FLOAT_SMALL));
 }
 
-inline double yaw(cocos2d::Vec3 point) {
+inline float yaw(cocos2d::Vec3 point) {
     return atan(distance(point.x, point.y) / (point.z + MATH_FLOAT_SMALL));
 }
 
-inline double roll(cocos2d::Vec3 point, cocos2d::Vec3 reference) {
+inline float roll(cocos2d::Vec3 point, cocos2d::Vec3 reference) {
     return roll(point) - roll(reference);
 }
 
-inline double pitch(cocos2d::Vec3 point, cocos2d::Vec3 reference) {
+inline float pitch(cocos2d::Vec3 point, cocos2d::Vec3 reference) {
     return pitch(point) - pitch(reference);
 }
 
-inline double yaw(cocos2d::Vec3 point, cocos2d::Vec3 reference) {
+inline float yaw(cocos2d::Vec3 point, cocos2d::Vec3 reference) {
     return yaw(point) - yaw(reference);
 }
